+++ int32_test.go
@@ -117,7 +117,7 @@ func TestInt32ScanTypeError(t *testing.T) {
 func TestInt32ValueInt(t *testing.T) {
 	val := Int32{Int32: 1, Valid: true}
 	got, err := val.Value()
-	if got != int32(1) || err != nil {
+	if got != int64(1) || err != nil {
 		t.Fatalf("want %v, but %v:", "1", got)
 	}
 }
@@ -125,7 +125,7 @@ func TestInt32ValueInt(t *testing.T) {
 func TestInt32ValueZero(t *testing.T) {
 	val := Int32{Int32: 0, Valid: true}
 	got, err := val.Value()
-	if got != int32(0) || err != nil {
+	if got != int64(0) || err != nil {
 		t.Fatalf("want %v, but %v:", 0, got)
 	}
 }
